# Conflicts:
#	bin/bsm.exe
#	bin/bsm_d.exe
#	bin/log1.txt
#	pro/BeamSplitting.pro.user
#	src/Beam.cpp
#	src/Handler.cpp
#	src/TracerBackScatterPoint.cpp
#	src/geometry/Polygon.h
#	src/scattering/Scattering.cpp
#	src/scattering/ScatteringConvex.cpp
#	src/scattering/ScatteringNonConvex.cpp

+++ src/particle/BulletRosette.cpp
@@ -81,7 +81,7 @@ BulletRosette::BulletRosette(const complex &refrIndex, double diameter,
 	SetDefaultCenters();
 	Reset();
 
-	for (int i = 0; i < nFacets; ++i)
+	for (size_t i = 0; i < nFacets; ++i)
 	{
 		defaultFacets[i].isVisibleIn = false;
 		defaultFacets[i].isVisibleOut = false;
